refactor: put interfaces in packge
