+++ data/src/main/java/com/madrid/data/dataSource/local/SearchLocalDataSource.kt
@@ -5,7 +5,7 @@ import com.madrid.data.dataSource.local.entity.MovieCategoryEntity
 import com.madrid.data.dataSource.local.entity.MovieEntity
 import com.madrid.data.dataSource.local.entity.RecentSearchEntity
 import com.madrid.data.dataSource.local.entity.SeriesEntity
-import com.madrid.data.repositories.SearchLocalSource
+import com.madrid.data.repositories.datasource.SearchLocalSource
 
 class SearchLocalDataSource(
     private val database: MovioDatabase
